+++ instrumentation-core/spring/spring-boot-autoconfigure/src/main/java/io/opentelemetry/instrumentation/spring/autoconfigure/webmvc/WebMvcFilterAutoConfiguration.java
@@ -16,7 +16,7 @@
 
 package io.opentelemetry.instrumentation.spring.autoconfigure.webmvc;
 
-import io.opentelemetry.instrumentation.springwebmvc.WebMVCTracingFilter;
+import io.opentelemetry.instrumentation.spring.webmvc.WebMvcTracingFilter;
 import io.opentelemetry.trace.Tracer;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -30,10 +30,10 @@ import org.springframework.web.filter.OncePerRequestFilter;
 @EnableConfigurationProperties(WebMVCProperties.class)
 @ConditionalOnProperty(prefix = "opentelemetry.trace.web", name = "enabled", matchIfMissing = true)
 @ConditionalOnClass(OncePerRequestFilter.class)
-public class WebMVCFilterAutoConfiguration {
+public class WebMvcFilterAutoConfiguration {
 
   @Bean
-  public WebMVCTracingFilter otelWebMVCTracingFilter(final Tracer tracer) {
-    return new WebMVCTracingFilter(tracer);
+  public WebMvcTracingFilter otelWebMvcTracingFilter(final Tracer tracer) {
+    return new WebMvcTracingFilter(tracer);
   }
 }
