+++ instrumentation-core/spring/spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -6,5 +6,5 @@ io.opentelemetry.instrumentation.spring.autoconfigure.exporters.logging.LoggingS
 io.opentelemetry.instrumentation.spring.autoconfigure.TracerAutoConfiguration,\
 io.opentelemetry.instrumentation.spring.autoconfigure.httpclients.resttemplate.RestTemplateAutoConfiguration,\
 io.opentelemetry.instrumentation.spring.autoconfigure.httpclients.webclient.WebClientAutoConfiguration,\
-io.opentelemetry.instrumentation.spring.autoconfigure.webmvc.WebMVCFilterAutoConfiguration,\
+io.opentelemetry.instrumentation.spring.autoconfigure.webmvc.WebMvcFilterAutoConfiguration,\
 io.opentelemetry.instrumentation.spring.autoconfigure.aspects.TraceAspectAutoConfiguration
